PDCL-6009 Directory rename, more informative error, regex test simplification.
