+++ src/view/components/formikReactSpectrum3/checkbox.styl
